pyproject.toml
[tool.poetry]
name = "weatherwatch"
version = "0.1.0"
description = ""
authors = ["Hao Tai <www991024@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.34"
databases = "^0.9.0"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.9.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
pydantic = "^2.9.1"
email-validator = "^2.2.0"
pydantic-settings = "^2.5.2"
asyncpg = "^0.29.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
bcrypt = "^4.2.0"
requests = "^2.32.3"
line-bot-sdk = "^3.13.0"
gunicorn = "^23.0.0"
flask = "^3.0.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
